backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
SQLAlchemy==2.0.36
PyMySQL==1.1.1
python-dotenv==1.0.1
requests==2.32.3
openpyxl==3.1.5
python-docx==1.1.2
docx2pdf==0.1.8
